% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_read_out.R
\name{bk_read_out}
\alias{bk_read_out}
\title{read in the data from outlet tables}
\usage{
bk_read_out(directory)
}
\arguments{
\item{directory}{The location of the folder containing out tables}
}
\value{
a dataframe containing data from the out tables
}
\description{
read in the data from outlet tables
}
\details{
The outlet files contain information about public library service outlets (central,  branch,  bookmobile,  and  books-by-mail-only  outlets)
}
