% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_read_pld.R
\name{bk_read_pld}
\alias{bk_read_pld}
\title{read in the data from pld tables}
\usage{
bk_read_pld(directory)
}
\arguments{
\item{directory}{The location of the folder containing pld tables}
}
\value{
a dataframe containing data from the pld tables
}
\description{
read in the data from pld tables
}
