% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_read_state.R
\name{bk_read_state}
\alias{bk_read_state}
\title{read in the data from state tables}
\usage{
bk_read_state(directory)
}
\arguments{
\item{directory}{The location of the folder containing state tables}
}
\value{
a dataframe containing data from the state tables
}
\description{
read in the data from state tables
}
